+++ CHANGELOG.md
@@ -2,6 +2,7 @@
 
 ## 3.1.0 
 - added new feature inactive
+- autoInit on documentReady
 
 ## 3.0.0 Early Bird
 - library custom builds with missing method warnings